@@ -4,11 +4,9 @@ Rails.application.routes.draw do
   mount Authentify::Engine => '/authentify'
   mount Commonx::Engine => '/common'
   
-  resource :session
-  
-  root :to => "authentify::sessions#new"
-  match '/signin',  :to => 'authentify::sessions#new'
-  match '/signout', :to => 'authentify::sessions#destroy'
-  match '/user_menus', :to => 'user_menus#index'
-  match '/view_handler', :to => 'authentify::application#view_handler'
+  root :to => "sessions#new", controller: :authentify
+  get '/signin',  :to => 'sessions#new', controller: :authentify
+  get '/signout', :to => 'sessions#destroy', controller: :authentify
+  get '/user_menus', :to => 'user_menus#index', controller: :main_app
+  get '/view_handler', :to => 'application#view_handler', controller: :authentify
 end
