@@ -50,10 +50,10 @@ module Dummy
     config.active_record.whitelist_attributes = true
 
     # Enable the asset pipeline
-    config.assets.enabled = true
+    #config.assets.enabled = true
 
     # Version of your assets, change this if you want to expire all your assets
-    config.assets.version = '1.0'
+    #config.assets.version = '1.0'
   end
 end
 
